se agregaron validaciones al user.service

+++ src/users/users.service.ts
@@ -1,9 +1,10 @@
-import { Injectable } from '@nestjs/common';
+import { HttpStatus, Injectable } from '@nestjs/common';
 import {InjectRepository} from  '@nestjs/typeorm'
 import {Repository} from 'typeorm';
 import { CreateUserDto } from './dto/create-user.dto';
 import { UpdateuserDto } from './dto/update-user.dto';
 import { User } from './user.entity';
+import { HttpException } from '@nestjs/common';
 
 @Injectable()
 export class UsersService {
@@ -11,8 +12,14 @@ export class UsersService {
 
     constructor(@InjectRepository(User) private userRepository: Repository<User>) {    
     }
-  
-    async createUser (user: CreateUserDto): Promise<CreateUserDto>{
+    
+    async createUser (user: CreateUserDto): Promise<any>{
+      const userFound = await this.userRepository.findOne({
+        where: {username: user.username}
+      });
+      if(userFound){
+        return new HttpException('User already exist', HttpStatus.CONFLICT)
+      }
       const newUser = this.userRepository.create(user);
       return await this.userRepository.save(newUser);
     }
@@ -20,16 +27,32 @@ export class UsersService {
    getUsers(){
     return this.userRepository.find();
    }
-   getUser(id: number){
-      return this.userRepository.findOne({
+   async getUser(id: number){
+    
+      const userFound = await this.userRepository.findOne({
         where: { id }
       });
+      if(!userFound){
+        return new HttpException('User not found', HttpStatus.NOT_FOUND);
+      }
+      return userFound;
    }
-   deleteUser(id : number){
+
+   async deleteUser(id : number){
+    const userFound = await this.userRepository.findOne({where: {id}});
+    if(!userFound){
+      return new HttpException('user not found', HttpStatus.NOT_FOUND);
+    };
     return this.userRepository.delete({id});
    }
 
-   updateUser(id: number, user: UpdateuserDto) {
-        return this.userRepository.update({id}, user);
+   async updateUser(id: number, user: UpdateuserDto) {
+    const userFound = await this.userRepository.findOne({where: {id}});
+    if(!userFound){
+      return new HttpException('user not found', HttpStatus.NOT_FOUND);
+    };
+        const userUpdated = Object.assign(userFound, user);
+        return this.userRepository.save(userUpdated);
+        // return this.userRepository.update({id}, user);
    }
 }
